+++ tests/testthat/setup.R
@@ -1,4 +1,5 @@
 if (not_on_cran()) {
+  message("Running tests setup")
   options(mc.cores = 2)
   obs <- filter_by_availability(
     germany_covid19_delta_obs,
@@ -37,8 +38,9 @@ if (not_on_cran()) {
 
   forecast_wrapper <- forecast(
     obs,
-    fit = fv_sample,
+    fit = silent_fv_sample,
     strains = c(1, 2), likelihood = FALSE,
-    adapt_delta = 0.98, max_treedepth = 15, chains = 2
+    adapt_delta = 0.98, max_treedepth = 15, chains = 2,
+    refresh = 0, show_messages = FALSE
   )
 }