+++ tests/testthat/test-plot_posterior.R
@@ -11,5 +11,7 @@ test_that("Can plot results from the two strain model", {
   skip_on_cran()
   p <- plot_posterior(posterior2)
   expect_ggplot(p[[1]])
-  expect_named(p, c("cases", "log_cases", "voc", "growth", "rt"))
+  expect_named(
+    p, c("cases", "log_cases", "voc_frac", "voc_advantage", "growth", "rt")
+  )
 })