+++ NEWS.md
@@ -1,3 +1,23 @@
+# forecast.vocs 0.0.8
+
+## Features
+
+* Adds `summary` and `plot` methods for `forecast()` and `fc_tidy_posterior()` as interfaces to lower level functions.
+* Adds a new summary variable for the transmission advantage of the VOC to the output from
+ `fv_tidy_posterior()`.
+* Adds a new plotting function `plot_voc_advantage()` to extract and plot this.
+* Adds `plot_voc_advantage()` to `plot_posterior` as an additional output.
+* Updates all example data to include this new output.
+
+## Breaking changes
+
+* The variant fraction label has been renamed `voc_frac` from `voc` due to the new variant advantage label.
+* `plot_voc()` has been renamed `plot_voc_frac()` due to the new `plot_voc_advantage()` function.
+
+## Bug fixes
+
+* Fixed an issue where `voc_label` was not being passed by `forecast()` to lower level methods and hence new labels specifying the VOC were not being assigned.
+
 # forecast.vocs 0.0.7
 
 * Further generalised the `forecast` family to accept custom data and forecast extraction functions.